Create an empty longitudinal profile plot.

This function generates an empty longitudinal profile plot using the 'plot_ly' function from the 'plotly' package.

Usage

lg_profile_empty()

Value

An empty longitudinal profile plot with a specified title.

Examples

# Create an empty longitudinal profile plot
empty_plot <- lg_profile_empty()
empty_plot