Add hydrological Regions in a Bassin to an existing Leaflet Map

This function adds regions within a bassin to an existing Leaflet map.

Utilisation

map_add_regions_in_bassin(
  map,
  bassin_click = bassin_click,
  regions_data = region_hydro
)

Arguments

map

An existing Leaflet map to which regions will be added.

bassin_click

A vector containing information about the clicked bassin.

regions_data

A sf data frame containing information about regions within the bassin.

Valeur de retour

An updated Leaflet map with regions added.

Exemples

library(leaflet)
library(dplyr)
#> 
#> Attachement du package : ‘dplyr’
#> Les objets suivants sont masqués depuis ‘package:stats’:
#> 
#>     filter, lag
#> Les objets suivants sont masqués depuis ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(sf)
# Create init bassin map
my_map <- map_init_bassins(bassins_data = bassin_hydrographique)
#> Error: l'argument "id_logo_ign_remonterletemps" est manquant, avec aucune valeur par défaut

# simulate bassin selected
selected_bassin <- bassin_hydrographique

# get centroid coordinate (in shiny see leaflet mapid_shape_click)
centre <- sf::st_centroid(selected_bassin)
#> Warning: st_centroid assumes attributes are constant over geometries
centre_coord <- as.data.frame(st_coordinates(centre)) %>%
  rename("lng" = X,
         "lat" = Y)
# map region
map <- map_add_regions_in_bassin(map = my_map,
                                 bassin_click = centre_coord, regions_data = region_hydrographique)
#> Error in eval(expr, envir, enclos): objet 'my_map' introuvable
map
#> Error in eval(expr, envir, enclos): objet 'map' introuvable